This commit will fix a circular dependency between settings service and config.js
